todays learning

+++ FastAPi/TodoApp/routers/todos.py
@@ -5,6 +5,7 @@ from sqlalchemy.orm import Session
 from database import  SessionLocal
 from schema import TodoRequest
 from typing import Annotated  
+from .auth import get_Current_user
 
 
 app = APIRouter()
@@ -19,16 +20,21 @@ def get_db():
         db.close()
         
 db_dependency=Annotated[Session, Depends(get_db)]
+user_dependency=Annotated[dict,Depends(get_Current_user)]
         
 @app.get("/todos")
-async def read_all(db:db_dependency):
-    return db.query(Todos).all()
+async def read_all(user:user_dependency,db:db_dependency):
+    if user is None:
+        raise HTTPException(status_code=401,detail="authentication failed")
+    return db.query(Todos).filter(Todos.Owner_id==user.get("id")).all()
 
 
 
 @app.get("/todos/{todo_id}",status_code=status.HTTP_200_OK)
-async def read_todo(db:db_dependency,todo_id:int=Path(gt=0)):
-    todo_model=db.query(Todos).filter(Todos.id==todo_id).first()
+async def read_todo(user:user_dependency,db:db_dependency,todo_id:int=Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401,detail="authentication failed")
+    todo_model=db.query(Todos).filter(Todos.id==todo_id).filter(Todos.Owner_id==user.get("id")).first()
     if todo_model is not None:
         return todo_model
     raise HTTPException(status_code=404,detail="todo is not found")
@@ -36,8 +42,12 @@ async def read_todo(db:db_dependency,todo_id:int=Path(gt=0)):
 
 
 @app.post("/todo",status_code=status.HTTP_201_CREATED)
-async def create_todo(db:db_dependency,todo_request:TodoRequest):
-    todo_model=Todos(**todo_request.model_dump())
+async def create_todo(user:user_dependency,db:db_dependency,todo_request:TodoRequest):
+    
+    if user is None:
+        raise HTTPException(status_code=401,detail="authentication failed")
+    
+    todo_model = Todos(**todo_request.model_dump(), Owner_id=user.get("id"))
     
     db.add(todo_model)
     db.commit()
@@ -46,8 +56,11 @@ async def create_todo(db:db_dependency,todo_request:TodoRequest):
 
 
 @app.put("/todo/{todo_id}",status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(db:db_dependency,todo_request:TodoRequest,todo_id:int=Path(gt=0)):
-    todo_model=db.query(Todos).filter(Todos.id==todo_id).first()
+async def update_todo(user:user_dependency,db:db_dependency,todo_request:TodoRequest,todo_id:int=Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401,detail="authentication failed")
+    
+    todo_model=db.query(Todos).filter(Todos.id==todo_id).filter(Todos.Owner_id==user.get("id")).first()
     if todo_model is  None:
         raise HTTPException(status_code=404,detail="todo not found")
     
@@ -60,14 +73,18 @@ async def update_todo(db:db_dependency,todo_request:TodoRequest,todo_id:int=Path
     db.add(todo_model)
     db.commit()
     db.refresh(todo_model)
+    return todo_model
     
     
 @app.delete("/todo/{todo_id}",status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db:db_dependency,todo_id:int=Path(gt=0)):
-    todo_model=db.query(Todos).filter(Todos.id==todo_id).first()
+async def delete_todo(user:user_dependency,db:db_dependency,todo_id:int=Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401,detail="authentication failed")
+    
+    todo_model=db.query(Todos).filter(Todos.id==todo_id).filter(Todos.Owner_id==user.get("id")).first()
     if todo_model is None:
          raise HTTPException(status_code=404,detail="todo not found")
         
-    db.query(Todos).filter(Todos.id==todo_id).delete()
+    db.query(Todos).filter(Todos.id==todo_id).filter(Todos.Owner_id==user.get("id")).delete()
     
     db.commit()
